backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
motor
pydantic
email-validator
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
scikit-learn
xgboost
pandas
numpy
easyocr
pillow
pytest
httpx